renamed login package to api because it is not just doing login and register anymore
