+++ api/endpoints_test.go
@@ -1,4 +1,4 @@
-package login
+package api
 
 /// needs actual tests
