+++ api/middlewares.go
@@ -1,4 +1,4 @@
-package login
+package api
 
 import (
 	"time"
