+++ api/service_test.go
@@ -1,4 +1,4 @@
-package login
+package api
 
 import (
 	"encoding/json"
