+++ api/transport.go
@@ -1,4 +1,4 @@
-package login
+package api
 
 // transport.go contains the binding from endpoints to a concrete transport.
 // In our case we just use a REST-y HTTP transport.
